requirements.txt
discord.py
asyncio
gtts